updates to python script

+++ code/FP_growth_0.1.py
@@ -12,9 +12,9 @@ df = df.iloc[:, 2:]
 # if the relative abundance is above 0.1, indicate presence with 1
 df1 = df.where(df <= 0.001, 1)
 # if the relative abundance is below 0.1, indicate presence with 0
-df1 = df1.where(df1 >= 0.001, 0).astype(int)
+df1 = df1.where(df1 >= 0.001, 0)
 # remove OTUs that have 0 variance
-df2 = df1.loc[:, df1.var() != 0]
+df2 = df1.loc[:, df1.var() != 0.0].astype(int)
 
 # write out the processed data that has been converted, this can be used in the follwoing steps as well as uploaded
 # into the Orange GUI to test out and compare
